+++ lib/protected_router.js
@@ -6,6 +6,6 @@ const config = require('./config').value;
 router.use('/(.*)', auth(config.authentication));
 
 // RSSHub
-router.get('/rsshub/rss', require('./routes/rsshub/rss'));
+router.get('/rsshub/routes', require('./routes/rsshub/routes'));
 
 module.exports = router;
