+++ lib/router.js
@@ -25,7 +25,8 @@ router.get('/', require('./routes/index'));
 router.get('/test/:id', require('./routes/test'));
 
 // RSSHub
-router.get('/rsshub/rss', require('./routes/rsshub/rss'));
+router.get('/rsshub/rss', require('./routes/rsshub/routes')); // 弃用
+router.get('/rsshub/routes', require('./routes/rsshub/routes'));
 
 // 1draw
 router.get('/1draw/', require('./routes/1draw/index'));
